immediate_mode_emulator를 yukino에서 sora로 옮김 상당히 자주 쓰일거같아서 이렇게 바꿈

+++ yukino/texture_atlas.cc
@@ -4,19 +4,19 @@
 
 namespace yukino {;
 float TextureAtlasSegment::GetAtlasX() const {
-  float tex_width = parent->tex->tex_header.tex_width;
+  float tex_width = (float)parent->tex->tex_header.tex_width;
   return x / tex_width;
 }
 float TextureAtlasSegment::GetAtlasY() const {
-  float tex_height = parent->tex->tex_header.tex_height;
+  float tex_height = (float)parent->tex->tex_header.tex_height;
   return y / tex_height;
 }
 float TextureAtlasSegment::GetAtlasW() const {
-  float tex_width = parent->tex->tex_header.tex_width;
+  float tex_width = (float)parent->tex->tex_header.tex_width;
   return w / tex_width;
 }
 float TextureAtlasSegment::GetAtlasH() const {
-  float tex_height = parent->tex->tex_header.tex_height;
+  float tex_height = (float)parent->tex->tex_header.tex_height;
   return h / tex_height;
 }
